% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pvalue.R
\name{extract_pvalue}
\alias{extract_pvalue}
\title{Extract p-values from a string.}
\usage{
extract_pvalue(x, full = TRUE)
}
\arguments{
\item{x}{a character string containing a p-value}

\item{full}{include "p=" or just the number, Default: TRUE}
}
\value{
a character string
}
\description{
Given a character string, this function will extract the p-value
}
\examples{
\dontrun{
if(interactive()){
 }
}
}
\seealso{
\code{\link[stringr]{str_extract}}
}
