% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_mean.R
\name{full_mean}
\alias{full_mean}
\title{Compare two statements Mean ± SEM}
\usage{
full_mean(means, group, comparison_string = "vs")
}
\arguments{
\item{means}{character vector of means.}

\item{group}{grouping column (e.g. "treatment")}

\item{comparison_string}{what word/expression do you want to use to compare
the means?, Default: "vs"}
}
\value{
A string with the format "Mean ± SEM unit vs Mean ± SEM"
}
\description{
Create a sentence comparing two simple means (e.g. from simple_means)
}
\examples{
\dontrun{
if(interactive()){
 ## Basic example
 x <- rnorm(50 + rnorm(50))
 simple_mean(x, digits = 3)

 ## With units
 simple_mean(iris$Petal.Length, units = "cm")
 }
}
}
