% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ttest.R
\name{print_ttest}
\alias{print_ttest}
\title{Format t-test results}
\usage{
print_ttest(tidy_ttest, digits = 2, ci = TRUE, ptype = "full")
}
\arguments{
\item{tidy_ttest}{Tidy data-frame with t-test results from broom::tidy(t.test()).}

\item{digits}{Number of significant digits.}

\item{ci}{Should confidence intervals be printed? Default: TRUE.}

\item{ptype}{Should p-values be presented in full or with a cutoff?}
}
\value{
A character vector.
}
\description{
Taking tidy results from a t-test, this function will create a
string ready for you to use in your paper.
}
\examples{
\dontrun{
if(interactive()){

 }
}
}
