% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pvalue.R
\name{extract_pvalue}
\alias{extract_pvalue}
\title{Extract p-values from a string.}
\usage{
extract_pvalue(x, full = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{full}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Given a character string, this function will extract
}
\examples{
\dontrun{
if(interactive()){
 }
}
}
\seealso{
\code{\link[stringr]{str_extract}}
}
